Do not modify argv[0] if argc is 0
The `ape -` variant with no arguments provides no way to shim in argv[0]
so the loader had no way to tell the binary what its path was. This is a
long-standing bug that looks like it has existed for as long as `ape -`,
and old loaders would, I guess, just overwrite the sentinel NULL that is
by convention the beginning and the end of a zero-length argv.

From the vantage point of TryElf I can't see a really good solution so I
just check for argc of 0 and let things be weird if a binary is run like
that.

This is only relevant with older binaries (as of #1170, those that don't
set `EF_APE_MODERN`.)