Copy constructor makes small strings when it can
We create a small string whenever it will fit, instead of only when r is
small. This is safe because we always memcpy the whole blob when we call
reserve, and optimizes the case when r has been truncated.

This also reworks the test to follow the idiom adopted elsewhere re stl,
and adds a helper function to tell if a string is small based on data().