wip ctl type traits
I am a bit ambivalent about this. <type_traits> is actually not that bad
by STL standards, “only” 148 transitive includes for the whole file. The
split out variants under __type_traits are also mostly reasonable - e.g.
__type_traits/is_same.h only includes 4 other STL files. Many others ask
for more, but nothing is all that egregious.

I do think we can probably provide an 80–95% solution with a fraction of
the footprint in CTL by for example eliding direct struct definitions in
favor of the _t or _v variants when the latter are simpler to write, and
omitting the more obscure tests in favor of a common core. I don’t think
it’s worth splitting things out by file but it might be worth it to e.g.
have a type_traits.h and a type_traits_advanced.h or something.