Fix %[hl]s field length/precision on printf funcs
Cosmopolitan currently has a completely wrong understanding (according
to the C Standard) of how to handle precision and field length for
wide strings, which have to be handled not by using the
monospace display width but instead by using the outputted byte
width (i.e. the UTF-8 length of the wide string).

This patch corrects this, corrects comments documenting the previous
behavior, adds tests for the behavior and refactors tests that tested
for the previous behavior to correctly use the new correct behavior.