Enable STL-style enable_shared_from_this
This way of doing it is kind of lame in that it is not possible to refer
to weak_this during object construction but this is how the STL does it.
The other would be to detect if there is a constructor that takes a weak
reference followed by Args in make_shared, but this does the wrong thing
if someone wrote their code not expecting that behavior to exist.