Replace "comprised of" with better expressions
The expression "comprised of" is not currently considered grammatically
correct. Depending on the context, either replace it with "comprises" or
"comprising", or replace it by the more obvious and less controversial
"made up of".

This change does not affect any third-party code where additional local
delta may complicate further updates to newer upstream versions.